Error handing added to db script
NOTE: u64 not supported by postgres

+++ yt_downlaoder/yt_download_manager/src/lib.rs
@@ -11,7 +11,7 @@ pub mod models {
         pub url: String,
         pub thumbnail_url: String,
         pub video_id: String,
-        pub size: i32,
+        pub size: i64,
     }
     impl From<web::Json<FetchMeta>> for FetchMeta {
         fn from(fetched: web::Json<FetchMeta>) -> Self {
